Remove query call on all pages except post edit
When the meta box got registered, it did a query for the description.
This is unnecessary on all pages except post edit page.

One concern with this solution is, if there would be default values in
the metabox registration, the defaults wouldn't be taken into account
when getting the values. But because News SEO doesn't have any defaults,
this is not an issue.