CS/QA: remove superfluous fall-back function
This method falls through to the parent method if the parent exists - which it does since WPSEO `3.0`.

As we are a good two years down the road, this fall-back can be safely removed.
The method does not need to be deprecated as any function calls to it will fall through to the parent method, which is exactly as intended.

As a side-note: up to now, this method would have never fallen through to the parent as there is a typo in the `is_callable()` - `:` vs `::`.