CS/QA: use strict comparisons when comparing strings
Whenever a variable/function call output is compared to a string, one should *always* use a strict comparison as otherwise the results will become unpredictable.

If the value of the variable is a boolean, the hard-coded string will be cast to boolean before comparing, if the value of the variable is an integer, the string will be cast to integer, etc.

So when the variable is expected to be a string and a comparison is made against a hard-coded string, this should always be a strict comparison.