WPSEO_News_UnitTestCase: add helper method to test private methods
If there is no functional reason to change the visibility of `private` methods to `protected`, it shouldn't be changed just to make the methods testable.

To still be able to test those methods in isolation, this helper method can be used.

Note: the helper method needs PHP 5.3.2+, which means that those methods can *not* be tested on PHP 5.2.