CS/QA: Rename class files
As discussed with @herregroen.

Adjusts the file names to be in line with the rules as laid out in the `Yoast.Files.FileName` sniff.
Ref: Yoast/yoastcs#48

* Remove `class-` prefixes from the plugin file names.
* Make sure that the file names accurately reflect the class names (excluding the prefixes which have been set at `(yoast_)wpseo_news_`).

Where necessary, adjusted any `include`/`require`s to point to the new file names.

Includes adjusting the PHPUnit config file, removing the file name `prefix` attribute and adjusting the `suffix` attribute.