Composer/PHPUnit: be explicit about supported PHPUnit versions
The unit tests for the Yoast News plugin, in part, depend on the WordPress native unit test setup and loads the WP native `bootstrap.php` file to that end.

This means that the unit tests for the plugin are compatible with the same PHPUnit versions as the WP native unit tests, which at this time is `^3.6.12 || ^4.5 || ^5.7 || ^6.0 || ^7.0`.
While PHPUnit 3.x is supported for testing on PHP 5.2, it is not listed in the `composer.json` file as Composer itself has a minimum requirement of PHP 5.3, so listing it is a moot point.

As PHPUnit is now included in the Composer setup, the Travis script needs to be tweaked:
* Defer to the Travis native PHPUnit version when it is within our range for faster builds.
    All Travis PHP images ship with PHPUnit already installed.
    It's only for those images which ship with an incompatible PHPUnit version (PHP 7.2 and higher ship with PHPUnit 8.x), that we need to install and use the PHPUnit version declared in the `composer.json` file.