Add CSS pipeline for the new metabox tab section CSS
The CSS is only enqueued on `post.php` and `post-new.php`, since it is for the metabox only.

Side effect is the development build set to true as expected now (false for `release` and `artifact` only).
Because a development build creates a sourcemap it requires the `clean:build-assets-js` task to be run in `build:js`.
However the JS is not set up properly in this repo. So the clean only removes the `*.map` files from the JS folder.