Schema: fix failing test
This test has been failing for a while - since commit 460da5aa4903f831ca3ab0659708516ae48ba594 - and the failure was hidden by errors in the Travis script.

The change made in the above mentioned commit means that the URL for `copyrightHolder` will always have a trailing slash before the hash `#`.
The unit tests were not updated for this change, causing them to fail.