Fix more failing tests
These tests have been failing for a while (if not forever) and the failure was hidden by errors in the Travis script.

The short of it is that these tests are using more modern PHPUnit syntax, while the tests are also still being run on older PHPUnit versions where that syntax was not yet available.