Tests: mark two tests for skipping on low PHP versions
These two tests both use the `DateTimeZone` class combined with the output of the `WPSEO_News_Sitemap_Timezone::wp_get_timezone_string()` method.

This method _may_ return the timezone as an UTC-offset, but the PHP `DateTimeZone` class did not support timezone offsets as input until PHP 5.5.10.

As support for PHP < 5.6 is about to be dropped, for now, let's just skip these tests on low PHP versions.

Ref: https://www.php.net/manual/en/datetimezone.construct.php#refsect1-datetimezone.construct-changelog