Travis: fix broken script
The Travis builds have been failing for quite a long time, but due to the script being set up incorrectly, this has gone (nearly) unnoticed.

The main problems were:
* Incorrect indentation.
* Incorrect folding.
* Commands where the exit code should determine build pass/fail being nested.
    Travis does not handle nested conditions well.

The above has all been cleaned up and the build is now correctly failing on all builds which run unit tests.

In addition to this:
* Removing `sudo`.
    Travis has not supported `sudo` for quite a while.
* Moving the merging of the code coverage test output to the `after_script` section.

Fixes 402