Tests: fix test double classes not being found
Some of the build failures are due to the test double classes not being found when they are supposed to be auto-loaded.

```
Fatal error: Class 'WPSEO_News_Sitemap_Item_Double' not found in /tmp/wordpress/src/wp-content/plugins/wpseo-news/integration-tests/sitemap-item-test.php on line 34
```

The problem here is that the `composer.json` contains an `autoload` `classmap` directive as well as an `autoload-dev` `classmap`.

In the Travis script, the `composer install` command is run with `--no-dev` on PHP < 7.2.
That means that for those PHP versions, the classmap generated did not contain the test classes, leading to the above mentioned kind of failures.

So, while using the Travis standard PHPUnit version in combination with a `--no-dev` `composer install` will lead to faster builds, we do need to make sure the `dev` classmap is generated, so it includes the unit test related files for autoloading.

Running `composer dump-autoload` (without `--no-dev`) after the installs are done, will fix this.