Bugfix: use the magic constant
The constant `__METHOD` does not exist. The magic PHP constant `__METHOD__` does.

Ref: https://www.php.net/manual/en/language.constants.predefined.php

Using a non-existent constant will cause PHP < 8 to thrown a warning + then treat the constant as a string.
In this case, that means that the deprecation warning would not list the method name of the deprecated method, but would say something like `__METHOD is deprecated since version ...`.