Build/PHPCS: temporary tweak for short array format
WordPressCS will demand long arrays and forbid the use of short arrays as of WPCS 2.2.0.
See: https://make.wordpress.org/core/2019/07/12/php-coding-standards-changes/

For YoastCS, however, a choice has been made to demand short arrays.

As support for PHP < 5.6 has been dropped, this can be enforced everywhere now and will be as of YoastCS 2.0.0 which is expected soon.

For now, just ignore use of short arrays.