Tests: use `expectOutputContains()`
WP Test Utils introduces a new `expectOutputContains()` method to complement the PHPUnit native `expectOutput*()` methods as it appears testing whether output contains a certain text string is a common pattern in the Yoast tests.

By default this method will ignore differences in line endings to improve the cross-OS compatibility of the tests.

This switches out the custom code to do the same in this test suite in favour of using the `expectOutputContains()` method instead.