Travis: fix integration test build on PHP >= 8.0
Composer 2.2 contains a a bug/over-optimization, which causes the install with changed, but compatible requirements to error out on "conflicting" requirements.

This commit puts a work-around for this in place.

It is recommended to revert this commit if/when the Composer bug has been fixed and a new version has been released.

Ref:
* Bug report: https://github.com/composer/composer/issues/10394
* Probable cause: https://github.com/composer/composer/pull/9620
* Similar fix in YoastSEO: https://github.com/Yoast/wordpress-seo/commit/b3999428193d5cdcdd0758dcb7dc060e6992c507